Add fix-path option.
On OS X electron frequently does not seem to pick up the PATH setting
correctly when launching Atom from a GUI, for example the Dock. This
means that shellcheck can not be found. One workaround is to specify the
absolute path to shellcheck, but this can be done automatically. This
provides a much better experience for users on OS X.

This should also resolve issue #7 for many users.

- Adds `enableFixPath` option.
- Adds documentation for option to README.
- Adds observer for option, once turned on it can't be reversed.
- Defaults `enableFixPath` to true for 'darwin` (OS X) platform.
- Provides user documentation for option, indicating its use in OS X.
- Adds `fix-path` dependency.